+++ CHANGELOG.md
@@ -7,9 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add Utils for languages
+
 ### Changed
 
+- Improve load of navbar
 - Update postcss config
+
+### Fixed
+
+- Fix undefined language
 
 ## [1.0.0] - 2023-04-06